requirements.txt
appdirs==1.4.4
asn1crypto==1.4.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
conda==4.3.16
Cython==0.29.22
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.4.dev0
Markdown @ file:///buildstream/freedesktop-sdk/components/python3-markdown.bst
MarkupSafe==1.1.1
meson @ file:///buildstream/freedesktop-sdk/components/meson.bst
pg8000==1.19.4
pipenv==2020.8.13
psycopg2-binary==2.8.6
pycosat==0.6.3
pydantic==1.7.3
Pygments==2.8.0.dev20210403
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scramp==1.4.0
six==1.15.0
SQLAlchemy==1.3.20
testing.common.database==2.0.3
testing.postgresql==1.3.0
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
Werkzeug==1.0.1
